(bugfix: #291): Add mobile version of EventCard
