+++ frontend/components/Calendar/BigCalendarDay.tsx
@@ -2,7 +2,7 @@ import { Box, Paper, Typography } from '@mui/material';
 
 export default function BigCalendarDay({ day }: { day: number }) {
   return (
-    <Paper style={{ width: '5rem', height: '5rem' }}>
+    <Paper style={{ minWidth: '5rem', width: '5rem', height: '5rem' }}>
       <Box
         height="1.25rem"
         style={{
