+++ frontend/components/News/articlestyles.ts
@@ -8,7 +8,7 @@ export const articleStyles = makeStyles((theme: Theme) => ({
     padding: "20px",
     borderRadius: "10px",
     marginBottom: "10px",
-    'a': {
+    a: {
       color: "#454545",
       textDecoration: "none",
     },
